Ecl:Ptzp: Pattern class command update
Data getting from user on pattern API was an error.
if the data type is float or double type , this was a mistake.
Because integer converter used. This problem solved.
Now, Float converters and variables are using for pattern

+++ ptzpatterninterface.h
@@ -7,7 +7,7 @@ class PtzPatternInterface
 {
 public:
 	virtual void positionUpdate(int pan, int tilt, int zoom) = 0;
-	virtual void commandUpdate(int pan, int tilt, int zoom, int c, int par1, int par2) = 0;
+	virtual void commandUpdate(int pan, int tilt, int zoom, int c, float par1, float par2) = 0;
 	virtual int start(int pan, int tilt, int zoom) = 0;
 	virtual void stop(int pan, int tilt, int zoom) = 0;
